Update README.md
This is a University project I produced. The project contains a comprehensive analysis of the software requirements needed for a system made to apply for grants for professors and their assistants from the faculties of our universities, to take scientific trips.

In this project I worked on a comprehensive analysis of software requirements that contains several projects formulated in this item.
1. To begin with, an analysis was made of a given document. According to the IEEE 29148:2011 standard for the STRUCTURE OF THE DOCUMENT, 2 positive aspects of the specification and 2 negative aspects of the specification are listed.
6 main problems in the TEXT of the document were identified and changes were suggested.
The procedure that I followed during the inspection of the document is described.
2. In the next phase, a session was held with stakeholders for Brainstorming and prioritizing ideas, and the following things are covered in the document:
a. Introduction
A brief overview of the objectives of this document
b. Initial user requirements
Based on the requirements arising from the brainstorming session and additional sources of information. These requirements are revised, written well and clearly, duplication eliminated, made consistent but without changing the scope (that is, without removing essential features).
These requirements are prioritized according to their potential impact on system improvement regardless of cost or development challenges.
c. Discussion
A query filtering / cleaning strategy has been developed and discussed briefly (~1/2 page).
The request prioritization strategy is discussed briefly (~ 1/2 page).
d. References
Included is a list of references I used in the document (eg documents, manuals, websites, competing systems, other authors..)
3. At this stage, an individual interview with the interested party (stakeholder) is scheduled through. The stakeholder with whom we interviewed was simulated.
I talked with different stakeholders (Head of Financial Service of FINKI, Head of Legal Service and Professor). With prepared questions for each of them based on the following criteria.

Organization and readiness:
Have existing systems, if any, been evaluated?
Do you know the results of the brainstorming session?
Are roles such as interviewer and scribe clearly identifiable?
Prepared a suitable list of questions?
Taking notes at the interview?
Running out of questions quickly?
Is it on track to meet the purpose of the interview?
Effectiveness:

How good are the questions and follow-up questions after the answers?
Can stakeholder objectives be identified?
Can the ideal functionalities be identified? The important ones? The dysfunctional aspects?
Conduct / Clarity
Are the questions clear and well structured?
Does the team use technical terms or assume that the customer has software engineering knowledge?
Courtesy
Courtesy during and after the interview
Accuracy
4. An elicitation document was developed based on the brainstorming session, the stakeholder interview and the knowledge gained through research on the needs of the project. The elicitation document contains all comprehensive elicitation sections.
5.The goal in this deliverable was to write a Software Requirements Specification (SRS), which includes useful attributes and tracking information. The SRS is linked to the revised objectives in the elicitation document.

The domain model from the elicitation document is included in the document. Goals are also included because those requirements are related to these goals.

The IEEE830 document template that can be found on the Internet is used. At this stage there was no need for the validation section, but it will have to be included in the final document after the validation meeting.

At least 5 user cases,  UML diagrams for user cases with text description (or possibly sequence diagrams) are included.

Each of these user cases covers at least one defined purpose indicating that the document has tracking information.

Inspection techniques covered in the document quality audit and requirements class are used. Traceability using a tool of choice is also covered.
The purpose of the document is quality, not quantity.
6. In the last stage, the final specification of the software requirements was made and thus the complete project is completed.